app/requirements.txt
streamlit
PyPDF2
langchain
faiss-cpu
openai
tiktoken
langchain_community
python-dotenv